pip-requirements.txt
absl-py==2.1.0
attrs==24.3.0
audioread @ file:///D:/bld/audioread_1725357575836/work
Brotli @ file:///D:/bld/brotli-split_1725267609074/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1734380492396/work/certifi
cffi @ file:///D:/bld/cffi_1725560792189/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
colorama==0.4.6
contourpy @ file:///C:/b/abs_853rfy8zse/croot/contourpy_1700583617587/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
dtaidistance==2.3.12
ffmpeg-python==0.2.0
flatbuffers==24.12.23
fonttools @ file:///D:/bld/fonttools_1735335831557/work
future==1.0.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1733298745555/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1733299205993/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1733298771451/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.36.1
imageio-ffmpeg==0.6.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1733223117029/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
jax==0.4.30
jaxlib==0.4.30
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1733736026804/work
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1733636780672/work
librosa @ file:///home/conda/feedstock_root/build_artifacts/librosa_1734618355842/work
llvmlite==0.43.0
matplotlib==3.9.2
mediapipe==0.10.20
ml_dtypes==0.5.1
moviepy==2.1.2
msgpack @ file:///D:/bld/msgpack-python_1725975146286/work
munkres==1.1.4
numba @ file:///D:/bld/numba_1718888166681/work
numpy @ file:///D:/bld/numpy_1707225561314/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=af5f40857bb7ceb2d3be562fa763863d2f954c104873e89b8ee370f75e047ecf
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0
opencv-python-headless==4.10.0
opt_einsum==3.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
ply @ file:///home/conda/feedstock_root/build_artifacts/ply_1733239724146/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1733421311631/work
proglog==0.1.10
protobuf==4.25.5
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1735698276679/work
PyQt5==5.15.9
PyQt5-sip @ file:///D:/bld/pyqt-split_1695417932202/work/pyqt_sip
PySocks @ file:///D:/bld/pysocks_1733217287171/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
scikit-learn @ file:///D:/bld/scikit-learn_1736496793625/work/dist/scikit_learn-1.6.1-cp39-cp39-win_amd64.whl#sha256=bb1e08460161bd7b8163df245d711be5c4d65ff892d330db2e71295806b53f85
scipy @ file:///C:/bld/scipy-split_1716470385682/work/dist/scipy-1.13.1-cp39-cp39-win_amd64.whl#sha256=490e922ef930c40e74290d95915a98295b9059d02849059317bab82ba7810581
sentencepiece==0.2.0
sip @ file:///D:/bld/sip_1697300548918/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sounddevice==0.5.1
soundfile @ file:///home/conda/feedstock_root/build_artifacts/pysoundfile_1734456627057/work
soxr @ file:///D:/bld/soxr-python_1725345578135/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1714400101435/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1733256695513/work
tornado @ file:///D:/bld/tornado_1732615926917/work
tqdm==4.67.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
unicodedata2 @ file:///D:/bld/unicodedata2_1736692548847/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
win_inet_pton @ file:///D:/bld/win_inet_pton_1733130564612/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
